Udacity Movational Meme/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.4
MarkupSafe==2.1.3
numpy==1.26.2
pandas==2.1.4
Pillow==10.1.0
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
